+++ lua/plugins/conform.lua
@@ -1,0 +1,10 @@
+-- lua/plugins/conform.lua
+
+return {
+    'stevearc/conform.nvim',
+    event = { 'BufWritePre' },
+    cmd   = { 'ConformInfo' },
+    config = function()
+        require('config.conform')
+    end,
+}