+++ lua/plugins/nvim-lint.lua
@@ -1,0 +1,9 @@
+-- lua/plugins/nvim-lint.lua
+
+return {
+    'mfussenegger/nvim-lint',
+    event = { 'BufReadPre', 'BufNewFile' },
+    config = function()
+        require('config.nvim-lint')
+    end,
+}